requirements.txt
Flask
flask-sqlalchemy
flask_login
flask_bcrypt
python-dotenv
flask-session
flask-mail
pyjwt
datetime
requests
psycopg2-binary
flask-migrate